Stylistic HTML tweaks and more...
Display a amount of cheats/codes a game has (defined in each game's
numOfCodes field).
Close MySQL connection when done with query.
A cheat's number, author credit, and note colums can be toggled in
config.php.
A code's author credit will read as "Unknown" if it isn't set in the
database.